flask_app/requirements.txt
flask
huggingface_hub
python-dotenv
llama_index
sentence-transformers
requests
streamlit